requirements.txt
pandas
datetime
gspread==4.0.1
oauth2client==4.1.3
google-auth==2.3.2
streamlit==0.86.0
gspread-dataframe==3.2.0